requirements.txt
# AWS Dependencies
boto3>=1.34.0

# Shared dependencies for all Generative AI Cloud Projects

# Core
python-dotenv==1.0.0

# Cloud SDKs
boto3==1.34.0              # AWS
azure-ai-ml==1.12.0        # Azure ML
azure-ai-openai==0.1.0     # Azure OpenAI
google-cloud-aiplatform==1.38.0  # GCP Vertex AI
openai==1.6.1              # OpenAI (fallback)

# Testing
pytest==7.4.4
pytest-cov==7.0.0
pytest-mock==3.15.1
google-cloud-run>=0.10.0

# General AI/ML Dependencies
openai>=1.12.0
langchain>=0.1.0
pinecone-client>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.1.0
numpy>=1.26.0
